requirements.txt
openai>=1.10.0
chromadb>=0.4.22
python-dotenv>=1.0.0
numpy>=1.24.0
sounddevice>=0.4.6
soundfile>=0.12.1 # Added soundfile for audio processing
pydub>=0.25.1 # For audio manipulation
d20==1.1.2 # Advanced dice parser and roller
ffmpeg-python>=0.2.0 # Required for pydub to work with various audio formats
PyPDF2>=3.0.0
pyaudio>=0.2.13
rich>=13.4.2
flask>=2.0.0
flask-socketio>=5.3.0
eventlet>=0.30.0
openai-agents[voice]>=0.0.8 # Use the latest available version
google-generativeai>=0.3.0 # For Gemini Pro 2 API access
